chore: update login journey steps due to flakyness in automated tests
