+++ web/tests/Web.E2ETests/Features/School/CreateComparators.feature
@@ -1,8 +1,8 @@
 Feature: School create comparator set
 
     Background:
-        Given I am on create comparators page for school with URN '777042'
-        And I have selected organisation '01: FBIT TEST - Community School (Open)' after logging in
+        Given I am on create comparators page for school with URN '777042'       
+        And I have signed in with organisation '01: FBIT TEST - Community School (Open)'
 
     Scenario: Can view create comparator set page
         When I click continue
