+++ web/tests/Web.E2ETests/Steps/School/CompareYourCostsSteps.cs
@@ -14,13 +14,6 @@ public class CompareYourCostsSteps(PageDriver driver)
     private IDownload? _download;
     private HomePage? _schoolHomePage;
 
-    [Given("I am not logged in")]
-    public async Task GivenIAmNotLoggedIn()
-    {
-        var page = await driver.Current;
-        await page.SignOut();
-    }
-
     [Given("I am on compare your costs page for school with URN '(.*)'")]
     public async Task GivenIAmOnCompareYourCostsPageForSchoolWithUrn(string urn)
     {
@@ -296,13 +289,6 @@ public class CompareYourCostsSteps(PageDriver driver)
     {
         Assert.NotNull(_comparisonPage);
         await _comparisonPage.AreComparisonChartsAndTablesDisplayed(false);
-    }
-
-    [Given("I have selected organisation '(.*)' after logging in")]
-    public async Task GivenIHaveSelectedOrganisationAfterLoggingIn(string organisation)
-    {
-        Assert.NotNull(_comparisonPage);
-        await _comparisonPage.SignIn(organisation);
     }
 
     [Given("I have created a custom comparator set for '(.*)' containing")]
